fix: complete GROUP 1 optimizations - consolidate utilities and fix test infrastructure
- Consolidate 7 utility files into 3 organized modules
- Fix all imports across 22+ files
- Update test infrastructure for new structure
- Add missing formatter functions
- Clean repository and remove deprecated files
