+++ src/main.py
@@ -12,7 +12,7 @@ import streamlit as st
 from src.db.migrations import run_migrations
 from src.ui.state.session_state import init_session_state
 from src.ui.styles.theme import load_theme
-from src.ui.utils.database_utils import render_database_health_widget
+from src.ui.utils.database_helpers import render_database_health_widget
 
 logger = logging.getLogger(__name__)
