+++ src/ui/pages/companies.py
@@ -11,11 +11,11 @@ import streamlit as st
 
 from src.services.company_service import CompanyService
 from src.ui.helpers.company_display import render_company_card_with_selection
-from src.ui.utils.session_helpers import (
+from src.ui.utils.database_helpers import (
     display_feedback_messages,
     init_session_state_keys,
 )
-from src.ui.utils.streamlit_context import is_streamlit_context
+from src.ui.utils.ui_helpers import is_streamlit_context
 
 logger = logging.getLogger(__name__)
