+++ src/ui/utils/__init__.py
@@ -4,7 +4,7 @@ This package contains utility modules for Streamlit UI functionality including
 background task management, data formatting, validation, and other helper functions.
 """
 
-from .background_tasks import (
+from .background_helpers import (
     BackgroundTaskManager,
     ProgressInfo,
     StreamlitTaskManager,
