--- src/ui/utils/background_tasks.py
@@ -1,534 +1,0 @@
-"""Streamlined background task management using Streamlit built-ins.
-
-This module provides a library-first approach to background task management,
-replacing complex custom threading with st.status() + simple threading for
-optimal performance and maintainability.
-
-Key improvements:
-- Uses st.status() for better UX and progress visualization
-- Simple threading instead of ThreadPoolExecutor
-- Direct st.session_state integration
-- Enhanced database session management for background threads
-- No memory leaks or cleanup needed
-- Real-time progress tracking with company-level details
-"""
-
-import logging
-import sys
-import threading
-import time
-import uuid
-
-from dataclasses import dataclass
-from datetime import datetime, timezone
-from typing import Any
-
-import streamlit as st
-
-from src.scraper import scrape_all
-from src.services.job_service import JobService
-from src.ui.utils.database_utils import (
-    clean_session_state,
-    suppress_sqlalchemy_warnings,
-)
-from src.ui.utils.validation_utils import safe_job_count
-
-logger = logging.getLogger(__name__)
-
-# Suppress SQLAlchemy warnings common in Streamlit context
-suppress_sqlalchemy_warnings()
-
-# Thread-safe lock for atomic session state operations
-_session_state_lock = threading.Lock()
-
-
-def _atomic_check_and_set(key: str, check_value: Any, set_value: Any) -> bool:
-    """Atomically check session state value and set new value if match.
-
-    Args:
-        key: Session state key to check/set
-        check_value: Expected current value to check for
-        set_value: New value to set if check passes
-
-    Returns:
-        True if check passed and value was set, False otherwise
-    """
-    with _session_state_lock:
-        current_value = st.session_state.get(key)
-        if current_value == check_value:
-            st.session_state[key] = set_value
-            return True
-        return False
-
-
-def _is_test_environment() -> bool:
-    """Detect if we're running in a test environment."""
-    return (
-        "pytest" in sys.modules
-        or "unittest" in sys.modules
-        or hasattr(st.session_state, "_test_mode")
-    )
-
-
-@dataclass
-class CompanyProgress:
-    """Individual company scraping progress tracking."""
-
-    name: str
-    status: str = "Pending"
-    jobs_found: int = 0
-    start_time: datetime | None = None
-    end_time: datetime | None = None
-    error: str | None = None
-
-
-@dataclass
-class ProgressInfo:
-    """Progress information for background tasks."""
-
-    progress: float
-    message: str
-    timestamp: datetime
-
-
-@dataclass
-class TaskInfo:
-    """Task information for background tasks."""
-
-    task_id: str
-    status: str
-    progress: float
-    message: str
-    timestamp: datetime
-
-
-class BackgroundTaskManager:
-    """Simple background task manager for compatibility."""
-
-    def __init__(self):
-        self.tasks = {}
-
-    def add_task(self, task_id: str, task_info: TaskInfo) -> None:
-        """Add a task to tracking."""
-        self.tasks[task_id] = task_info
-
-    def get_task(self, task_id: str) -> TaskInfo | None:
-        """Get task information."""
-        return self.tasks.get(task_id)
-
-    def remove_task(self, task_id: str) -> None:
-        """Remove a task from tracking."""
-        self.tasks.pop(task_id, None)
-
-
-class StreamlitTaskManager(BackgroundTaskManager):
-    """Streamlit-specific task manager."""
-
-
-def render_scraping_controls() -> None:
-    """Render scraping controls with progress tracking.
-
-    Uses library-first st.status() for progress visualization and
-    st.session_state for state management. Includes database session
-    cleanup to prevent contamination.
-    """
-    # Clean any contaminated database objects from session state
-    clean_session_state()
-
-    # Initialize scraping state and status to prevent UI flicker
-    if "scraping_active" not in st.session_state:
-        st.session_state.scraping_active = False
-    if "scraping_results" not in st.session_state:
-        st.session_state.scraping_results = None
-    if "scraping_status" not in st.session_state:
-        st.session_state.scraping_status = "Ready to start scraping..."
-
-    # Create persistent status container to prevent flicker
-    status_container = st.empty()
-    status_container.info(st.session_state.scraping_status)
-
-    col1, col2 = st.columns([1, 1])
-
-    with col1:
-        if not st.session_state.scraping_active and st.button(
-            "🔍 Start Scraping", type="primary"
-        ):
-            start_scraping()
-
-    with col2:
-        if st.session_state.scraping_active and st.button(
-            "⏹️ Stop Scraping", type="secondary"
-        ):
-            st.session_state.scraping_active = False
-            st.session_state.scraping_status = "Scraping stopped"
-            st.rerun()
-
-
-def start_scraping() -> None:
-    """Start scraping with synchronous progress initialization.
-
-    Uses library-first approach for better progress tracking.
-    """
-    logger.info("start_scraping called - using library-first st.status approach")
-
-    # Get active companies synchronously before any background processing
-    try:
-        active_companies = JobService.get_active_companies()
-        logger.info("Found %d active companies to scrape", len(active_companies))
-    except Exception as e:
-        st.error(f"❌ Failed to load companies: {e!s}")
-        logger.exception("Failed to load active companies")
-        return
-
-    # Handle empty companies list early
-    if not active_companies:
-        st.warning("⚠️ No active companies found to scrape")
-        logger.warning("No active companies found for scraping")
-        return
-
-    # Initialize session state immediately (synchronously)
-    st.session_state.scraping_active = True
-    st.session_state.scraping_status = "Initializing scraping..."
-
-    # Initialize company progress tracking BEFORE background task
-    st.session_state.company_progress = {}
-    for company_name in active_companies:
-        st.session_state.company_progress[company_name] = CompanyProgress(
-            name=company_name,
-            status="Pending",
-            jobs_found=0,
-            start_time=None,
-            end_time=None,
-        )
-
-    logger.info("Initialized progress for %d companies", len(active_companies))
-
-    # Force immediate UI update to show initialized progress
-    st.rerun()
-
-    def scraping_task():
-        """Background task with proper session state management."""
-        try:
-            logger.info("=== SCRAPING TASK STARTED ===")
-            start_time = datetime.now(timezone.utc)
-
-            # Update session state status for persistent display
-            st.session_state.scraping_status = "🔍 Scraping job listings..."
-
-            # Process companies sequentially with progress tracking
-            for i, company_name in enumerate(active_companies):
-                # Check if scraping was stopped
-                if not st.session_state.get("scraping_active", False):
-                    logger.info("Scraping stopped by user")
-                    return
-
-                # Mark current company as scraping
-                if company_name in st.session_state.company_progress:
-                    st.session_state.company_progress[company_name].status = "Scraping"
-                    st.session_state.company_progress[
-                        company_name
-                    ].start_time = datetime.now(timezone.utc)
-
-                    # Update overall progress
-                    progress_pct = (i + 0.5) / len(active_companies) * 100
-                    st.session_state.scraping_status = (
-                        f"📊 Scraping {company_name}... ({progress_pct:.0f}%)"
-                    )
-                    logger.info(
-                        "Processing company %d/%d: %s",
-                        i + 1,
-                        len(active_companies),
-                        company_name,
-                    )
-
-                # Add small delay to show progression (configurable for demo)
-                if not _is_test_environment():
-                    time.sleep(0.1)  # Reduced from 0.5s for better responsiveness
-
-            # Update status in session state (no UI in background thread)
-            st.session_state.scraping_status = "📊 Running scraper..."
-            logger.info("Background scraping workflow initialized")
-
-            # Get job limit from session state and validate it
-            from src.scraper_company_pages import DEFAULT_MAX_JOBS_PER_COMPANY
-
-            max_jobs_per_company = st.session_state.get(
-                "max_jobs_per_company", DEFAULT_MAX_JOBS_PER_COMPANY
-            )
-            try:
-                max_jobs_per_company = int(max_jobs_per_company)
-                if max_jobs_per_company < 1:
-                    max_jobs_per_company = DEFAULT_MAX_JOBS_PER_COMPANY
-            except (ValueError, TypeError):
-                max_jobs_per_company = DEFAULT_MAX_JOBS_PER_COMPANY
-
-            # Execute scraping (preserves existing scraper.py logic)
-            logger.info("Starting scrape_all with max_jobs=%d", max_jobs_per_company)
-            result = scrape_all(max_jobs_per_company)
-            logger.info("scrape_all completed with result: %s", result)
-
-            # Update company progress with real results
-            for company_name in active_companies:
-                if company_name in st.session_state.company_progress:
-                    company_progress = st.session_state.company_progress[company_name]
-                    company_progress.status = "Completed"
-                    company_progress.end_time = datetime.now(timezone.utc)
-
-                    # Set real job count from scraper results with type safety
-                    raw_job_count = result.get(company_name, 0)
-                    company_progress.jobs_found = safe_job_count(
-                        raw_job_count, company_name
-                    )
-
-                    # If start_time wasn't set, estimate it
-                    if company_progress.start_time is None:
-                        company_progress.start_time = start_time
-
-            # Show completion
-            total_jobs = sum(result.values()) if result else 0
-            completion_msg = (
-                f"✅ Scraping Complete! Found {total_jobs} jobs across "
-                f"{len(active_companies)} companies"
-            )
-
-            st.session_state.scraping_status = completion_msg
-            logger.info("Scraping completed: %s", completion_msg)
-
-            # Store results
-            st.session_state.scraping_results = result
-
-            # In test environments, check if we should stay active
-            if _is_test_environment() and st.session_state.get(
-                "_test_stay_active", False
-            ):
-                # Keep scraping_active = True for start/stop workflow tests
-                pass  # Don't set scraping_active = False
-            else:
-                st.session_state.scraping_active = False
-
-        except Exception as e:
-            error_msg = f"❌ Scraping failed: {e}"
-
-            # Mark any scraping companies as error with safe attribute access
-            company_dict = getattr(st.session_state, "company_progress", None)
-            if company_dict:
-                for company_progress in company_dict.values():
-                    if company_progress.status == "Scraping":
-                        company_progress.status = "Error"
-                        # Safe attribute assignment - error field exists in dataclass
-                        if hasattr(company_progress, "error"):
-                            company_progress.error = str(e)
-                        company_progress.end_time = datetime.now(timezone.utc)
-
-            st.session_state.scraping_status = error_msg
-
-            # In test environments, respect stay_active setting even on error
-            if _is_test_environment() and st.session_state.get(
-                "_test_stay_active", False
-            ):
-                # Keep scraping_active = True for start/stop workflow tests
-                pass  # Don't set scraping_active = False
-            else:
-                st.session_state.scraping_active = False
-            logger.exception("Scraping failed")
-
-    # In test environments, run synchronously to avoid threading issues
-    if _is_test_environment():
-        scraping_task()
-    else:
-        # Store thread reference for proper cleanup
-        logger.info("Creating background thread for scraping task")
-        thread = threading.Thread(target=scraping_task, daemon=False)
-        st.session_state.scraping_thread = thread
-        thread.start()
-        logger.info("Background thread started: %s", thread.is_alive())
-
-
-# Simple API functions (preserve compatibility)
-def is_scraping_active() -> bool:
-    """Check if scraping is currently active."""
-    return st.session_state.get("scraping_active", False)
-
-
-def get_scraping_results() -> dict[str, Any]:
-    """Get results from the last scraping operation."""
-    return st.session_state.get("scraping_results", {})
-
-
-# Compatibility functions for existing code
-def get_task_manager() -> StreamlitTaskManager:
-    """Get or create the task manager instance."""
-    if "task_manager" not in st.session_state:
-        st.session_state.task_manager = StreamlitTaskManager()
-    return st.session_state.task_manager
-
-
-def start_background_scraping(stay_active_in_tests: bool = False) -> str:
-    """Start background scraping using Streamlit native approach.
-
-    This function uses a simple session state flag to trigger scraping
-    instead of complex threading, avoiding database session conflicts.
-
-    Args:
-        stay_active_in_tests: If True, keeps scraping_active=True in test environments
-                             even after completion. Used for start/stop workflow tests.
-    """
-    task_id = str(uuid.uuid4())
-    logger.info("start_background_scraping called with task_id: %s", task_id)
-
-    # Initialize session state immediately
-    if "task_progress" not in st.session_state:
-        st.session_state.task_progress = {}
-
-    # Store task info in session state
-    st.session_state.task_progress[task_id] = ProgressInfo(
-        progress=0.0,
-        message="Starting scraping...",
-        timestamp=datetime.now(timezone.utc),
-    )
-    st.session_state.task_id = task_id
-
-    # Store test behavior preference
-    if _is_test_environment():
-        st.session_state._test_stay_active = stay_active_in_tests
-
-    # Set scraping trigger flag - this will be processed by @st.fragment
-    st.session_state.scraping_trigger = True
-    st.session_state.scraping_active = True
-    st.session_state.scraping_status = "Initializing scraping..."
-
-    logger.info("Scraping trigger set - will be processed by fragment")
-    return task_id
-
-
-def stop_all_scraping() -> int:
-    """Stop all scraping operations with proper thread cleanup."""
-    stopped_count = 0
-    if st.session_state.get("scraping_active", False):
-        st.session_state.scraping_active = False
-        st.session_state.scraping_status = "Scraping stopped"
-
-        # Clean up thread reference if exists
-        if hasattr(st.session_state, "scraping_thread"):
-            thread = st.session_state.scraping_thread
-            if thread and thread.is_alive():
-                # Thread will terminate when it checks scraping_active
-                thread.join(timeout=5.0)  # Wait up to 5 seconds
-            delattr(st.session_state, "scraping_thread")
-        stopped_count = 1
-    return stopped_count
-
-
-def get_scraping_progress() -> dict[str, ProgressInfo]:
-    """Get current scraping progress."""
-    return st.session_state.get("task_progress", {})
-
-
-def get_company_progress() -> dict[str, CompanyProgress]:
-    """Get current company-level scraping progress.
-
-    Returns:
-        Dictionary mapping company names to their CompanyProgress objects.
-    """
-    return st.session_state.get("company_progress", {})
-
-
-@st.fragment(run_every=1)  # Check every second for scraping trigger
-def process_scraping_trigger() -> None:
-    """Process scraping trigger using Streamlit native fragment approach.
-
-    This function runs every second and checks if scraping needs to be triggered.
-    Uses atomic check-and-set to prevent race conditions between fragments.
-    """
-    # Atomically check if trigger is set and reset it to prevent race conditions
-    if not _atomic_check_and_set("scraping_trigger", True, False):
-        return
-
-    logger.info("Processing scraping trigger - starting actual scraping")
-
-    try:
-        # Get active companies
-        active_companies = JobService.get_active_companies()
-        if not active_companies:
-            st.session_state.scraping_status = "No active companies found"
-            st.session_state.scraping_active = False
-            return
-
-        # Initialize company progress
-        if (
-            "company_progress" not in st.session_state
-            or st.session_state.company_progress is None
-        ):
-            st.session_state.company_progress = {}
-
-        for company_name in active_companies:
-            st.session_state.company_progress[company_name] = CompanyProgress(
-                name=company_name,
-                status="Pending",
-                jobs_found=0,
-                start_time=None,
-                end_time=None,
-            )
-
-        # Update status
-        st.session_state.scraping_status = "Running scraper..."
-
-        # Get job limit
-        from src.scraper_company_pages import DEFAULT_MAX_JOBS_PER_COMPANY
-
-        max_jobs_per_company = st.session_state.get(
-            "max_jobs_per_company", DEFAULT_MAX_JOBS_PER_COMPANY
-        )
-
-        # Execute scraping directly (no threading)
-        logger.info("Executing scrape_all with max_jobs=%d", max_jobs_per_company)
-        result = scrape_all(max_jobs_per_company)
-        logger.info("scrape_all completed with result: %s", result)
-
-        # Update company progress with results
-        for company_name in active_companies:
-            if company_name in st.session_state.company_progress:
-                company_progress = st.session_state.company_progress[company_name]
-                company_progress.status = "Completed"
-                company_progress.end_time = datetime.now(timezone.utc)
-                company_progress.start_time = (
-                    company_progress.start_time or datetime.now(timezone.utc)
-                )
-
-                # Set job count from results
-                raw_job_count = result.get(company_name, 0)
-                company_progress.jobs_found = safe_job_count(
-                    raw_job_count, company_name
-                )
-
-        # Update final status
-        total_jobs = sum(result.values()) if result else 0
-        st.session_state.scraping_status = (
-            f"✅ Scraping Complete! Found {total_jobs} jobs across "
-            f"{len(active_companies)} companies"
-        )
-        st.session_state.scraping_results = result
-
-        # Mark as inactive unless in test mode
-        if not (
-            _is_test_environment() and st.session_state.get("_test_stay_active", False)
-        ):
-            st.session_state.scraping_active = False
-
-    except Exception as e:
-        logger.exception("Scraping failed in fragment processor")
-        st.session_state.scraping_status = f"❌ Scraping failed: {e}"
-
-        # Mark companies as error
-        if "company_progress" in st.session_state and st.session_state.company_progress:
-            for company_progress in st.session_state.company_progress.values():
-                if company_progress.status in ["Pending", "Scraping"]:
-                    company_progress.status = "Error"
-                    company_progress.error = str(e)
-                    company_progress.end_time = datetime.now(timezone.utc)
-
-        # Mark as inactive unless in test mode
-        if not (
-            _is_test_environment() and st.session_state.get("_test_stay_active", False)
-        ):
-            st.session_state.scraping_active = False
